chose(Home): Rename for_you_slate_provider.py
